+++ test/test_helper.rb
@@ -5,8 +5,5 @@ $LOAD_PATH.unshift File.expand_path('../lib', __dir__)
 require 'robot_lab'
 require 'robot_lab/durable'
 
-require 'fileutils'
-require 'tmpdir'
-
 require 'minitest/autorun'
 require 'minitest/pride'